Refactor facility management and buff systems
- Introduced new BuffType enum for additional buff types.
- Consolidated buff application logic in CharacterBuffManager to handle various buffs more efficiently.
- Updated FacilityBase to use float for health points and damage calculations.
- Implemented new FacilitySaveData structure for saving facility states.
- Enhanced FacilityManager to manage buffs and damage calculations more effectively.
- Added WorkbenchSaveData structure for saving workbench states.
- Updated WorkbenchManager to handle new buff logic and save/load functionality.
- Refactored ThermalControl and PowerSupply facilities to utilize new buff system.
- Improved WorkbenchDialog to reflect changes in working time calculations.

+++ Assets/VoidSurvival_10-Days/Scripts/Facilities/FacilityBase.cs
@@ -26,8 +26,8 @@ public abstract class FacilityBase : MonoBehaviour
 
     protected const int MaxLevel = 4;
     public int Level = 1;
-    public int CurrentHealthPoint => _currentHealthPoint;
-    protected int _currentHealthPoint;
+    public float CurrentHealthPoint => _currentHealthPoint;
+    protected float _currentHealthPoint;
 
     #region Inspector Fields
     // インスペクターで設定する項目
@@ -65,7 +65,8 @@ public abstract class FacilityBase : MonoBehaviour
     /// 施設の初期化
     /// </summary>
     /// <param name="level"></param>
-    public void initialize(int level, int currentHealthPoint)
+    /// <param name="currentHealthPoint"></param>
+    public void initialize(int level, float currentHealthPoint)
     {
         Level = level;
         _currentHealthPoint = currentHealthPoint;
@@ -143,7 +144,7 @@ public abstract class FacilityBase : MonoBehaviour
     /// 施設にダメージを与える
     /// </summary>
     /// <param name="damageAmount"></param>
-    public virtual void DamageFacility(int damageAmount)
+    public virtual void DamageFacility(float damageAmount)
     {
         _currentHealthPoint -= damageAmount;
         if (_currentHealthPoint < 0)
